+++ test/flame/projects_test.exs
@@ -29,10 +29,10 @@ defmodule Flame.ProjectsTest do
                {:error, :duration_too_long}
     end
 
-    test "exchanges id token for a cookie session", %{user: user} do
-      {:ok, token} = Flame.Accounts.sign_in(user.local_id)
+    test "exchanges id token for a cookie session", %{user: user, password: password} do
+      {:ok, token} = Flame.Accounts.sign_in(user.email, password)
 
-      assert {:ok, _} = Projects.create_session_cookie(token.value, @duration)
+      assert {:ok, _} = Projects.create_session_cookie(token, @duration)
     end
   end
 
@@ -40,19 +40,26 @@ defmodule Flame.ProjectsTest do
     test "passes when token itself is valid" do
       mock_cookie =
         ExFirebaseAuth.Mock.generate_cookie("1234", %{
-          "email" => "foo@example.com",
           "iat" => Epoch.now(),
           "exp" => Epoch.now() + 10,
           "auth_time" => Epoch.now() - 10
         })
+        |> Flame.SessionCookie.new()
 
-      assert {:ok, %Flame.Token{sub: "1234", email: "foo@example.com"}} =
-               Projects.verify_session(mock_cookie)
+      assert {:ok, %Flame.SessionCookie{sub: "1234"}} = Projects.verify_session(mock_cookie)
     end
 
     test "fails on expired tokens" do
+      now = Epoch.now()
+
       mock_cookie =
-        ExFirebaseAuth.Mock.generate_cookie("user_id", %{"email" => "foo@example.com", "exp" => 1})
+        ExFirebaseAuth.Mock.generate_cookie("user_id", %{
+          "email" => "foo@example.com",
+          "exp" => now - 10,
+          "iat" => now - 20,
+          "auth_time" => now - 20
+        })
+        |> Flame.SessionCookie.new()
 
       assert Projects.verify_session(mock_cookie) == {:error, "Expired JWT"}
     end
@@ -67,6 +74,7 @@ defmodule Flame.ProjectsTest do
           "exp" => Epoch.now() + 10,
           "auth_time" => Epoch.now() - 10
         })
+        |> Flame.SessionCookie.new()
 
       # NOTE kid is missing in emulator, so using bypass
       bypass = Bypass.open(port: 3000)
@@ -80,7 +88,7 @@ defmodule Flame.ProjectsTest do
       data = %{users: [user_fixture()]}
       mock_response(bypass, "lookup", data, 200)
 
-      assert {:ok, %Flame.Token{}} = Projects.verify_session(cookie, verify: true)
+      assert {:ok, %Flame.SessionCookie{}} = Projects.verify_session(cookie, verify: true)
     after
       Application.put_env(:flame, Flame, @existing_config)
     end
@@ -95,6 +103,7 @@ defmodule Flame.ProjectsTest do
           "exp" => now + 10,
           "auth_time" => now - 10
         })
+        |> Flame.SessionCookie.new()
 
       # NOTE kid is missing in emulator, so using bypass
       bypass = Bypass.open(port: 3000)
@@ -116,6 +125,7 @@ defmodule Flame.ProjectsTest do
     test "fails on expired tokens" do
       mock_cookie =
         ExFirebaseAuth.Mock.generate_cookie("user_id", %{"email" => "foo@example.com", "exp" => 1})
+        |> Flame.SessionCookie.new()
 
       assert Projects.verify_session(mock_cookie, verify: true) == {:error, "Expired JWT"}
     end
@@ -130,6 +140,7 @@ defmodule Flame.ProjectsTest do
           "exp" => now + 10,
           "auth_time" => now - 10
         })
+        |> Flame.SessionCookie.new()
 
       # NOTE kid is missing in emulator, so using bypass
       bypass = Bypass.open(port: 3000)
@@ -157,6 +168,7 @@ defmodule Flame.ProjectsTest do
           "iat" => now,
           "auth_time" => now - 10
         })
+        |> Flame.SessionCookie.new()
 
       assert Projects.verify_session(mock_cookie, verify: true) == {:error, :user_not_found}
     end