+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Change API to return `{:ok, Flame.Token.t()}` instead of `{:ok, String.t(), String.t()}`
+- Change API to return `{:ok, Flame.IdToken.t() | Flame.SessionCookie.t()}` instead of `{:ok, String.t(), String.t()}`
 - Remove Flame.Finch from supervisor, use built-in Finch from Goth for requests.
 
 ## [0.1.0] - 2022-04-28